Moved models out of shared folder
