+++ src/client/models/Entity.ts
