+++ src/client/models/NotificationEvent.ts
