+++ src/provider/models/Result.ts
