+++ src/provider/persistence/dataLayer/IDatastore.ts
@@ -1,4 +1,4 @@
-import {Entity} from '../../../shared/models/Entity';
+import {Entity} from '../../../client/models/Entity';
 import {ITable} from '../models/ITable';
 import {PageInfo} from '../models/PageInfo';
