+++ src/provider/persistence/dataLayer/repositories/RepositoryFactory.ts
@@ -1,6 +1,6 @@
-import {Entity} from '../../../../shared/models/Entity';
-import {Notification} from '../../../../shared/models/Notification';
-import {Settings} from '../../../../shared/models/Settings';
+import {Entity} from '../../../../client/models/Entity';
+import {Notification} from '../../../../client/models/Notification';
+import {Settings} from '../../../models/Settings';
 import {ISenderInfo} from '../../../models/ISenderInfo';
 import {IDatastore} from '../IDatastore';
 import {IndexedDb} from '../IndexedDb';
