+++ src/provider/ui/models/toast/IToast.ts
@@ -1,4 +1,4 @@
-import {Notification} from '../../../../shared/models/Notification';
+import {Notification} from '../../../../client/models/Notification';
 import {ISenderInfo} from '../../../models/ISenderInfo';
 
 export interface IToast {
