+++ src/provider/ui/toast/index.tsx
@@ -6,7 +6,7 @@ import { ISenderInfo } from '../../models/ISenderInfo';
 import { CHANNEL_NAME } from '../../../shared/config';
 import { NotificationCenterAPI, NotificationCenterEventMap } from '../NotificationCenterAPI';
 import { ProviderIdentity } from 'openfin/_v2/api/interappbus/channel/channel';
-import { Notification } from '../../../shared/models/Notification';
+import { Notification } from '../../../client/models/Notification';
 
 declare var window: Window&{openfin: {notifications: NotificationCenterAPI}};
