--- src/shared/models/Application.ts
@@ -1,6 +1,0 @@
-export interface Application {
-    name: string;
-    uuid: string;
-    remote: boolean;
-    blocked: boolean;
-}
