--- src/shared/models/Centre.ts
@@ -1,4 +1,0 @@
-export interface Centre {
-    order: string;
-    history: number;
-}
